services/scanner/requirements.txt
alembic==1.13.2
asyncpg==0.29.0
bcrypt==4.1.2
email-validator==2.2.0
fastapi==0.115.0
httpx==0.27.2
openai==1.75.0
passlib==1.7.4
pydantic-settings==2.5.2
pytest==8.3.3
reportlab==4.2.5
sqlalchemy==2.0.35
uvicorn[standard]==0.30.6
